+++ views/settings/profile.php
@@ -13,7 +13,6 @@
 use dektrium\user\helpers\Timezone;
 use kartik\widgets\DatePicker;
 use kartik\widgets\FileInput;
-use kartik\widgets\Select2;
 use yii\bootstrap\ActiveForm;
 use yii\helpers\ArrayHelper;
 use yii\helpers\Html;
@@ -121,8 +120,14 @@ $this->params['breadcrumbs'][] = $this->title;
                     } ?>
 
                     <?php if(Yii::$app->getModule('userextended')->account) {
-                        echo $form->field($model, 'birthday')->widget(Select2::class, [
+                        echo $form->field($model, 'account')->textInput([
+                            'placeholder' => Yii::t('traits', 'Account')
+                        ]);
+                    } ?>
 
+                    <?php if(Yii::$app->getModule('userextended')->contact) {
+                        echo $form->field($model, 'contact')->textInput([
+                            'placeholder' => Yii::t('traits', 'Contact')
                         ]);
                     } ?>